+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Add Suntory as build Makefile
+
 ## 2.5.0 - 2025-11-17
 ### Changed
 - Set min node engine to >= 20.0.0